Scores need rounding for precision and scale.

+++ app/models/score.rb
@@ -3,7 +3,7 @@ class Score
   property :id,         DataMapper::Types::Digest::SHA1.new(:source, :url), :key => true, :nullable => false
   property :source,     String, :nullable => false
   property :url,        URI, :length => 255, :nullable => false
-  property :score,      Float, :default => 0
+  property :score,      Integer, :default => 0
   property :created_at, DateTime
   property :updated_at, DateTime
